requirements.txt
Flask
python-dotenv
pyodbc
azure-storage-blob
requests
opencensus-ext-azure
gunicorn


